requirements.txt
cohere>=4.9.0
qdrant-client>=1.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.4.0
certifi>=2023.0.0